feat(): Print sonar, json and string report to file
This feature allows for the integration of find_unused_sass_variable
to a sonarqube instance. Therefore, it introduces the --output-file
option to print reports of unused variabes to an output file and
the --formatter option to set the output file's format.

Supported formats include string (default), json and sonar.

The sonar format creates a JSON report file in sonarqube's generic issue
format to being imported into sonarqube.  Ununsed variabels are
displayed with a severity of minor, as a code smell and with file and
textRange information for the primaryLocation property.
See https://docs.sonarqube.org/7.9/analysis/generic-issue/ for further
information.

The string format corresponds to the usual format of the console output.

The json format outpus a json file containing each unused variable
with there file, columnstart, -end and linestart and -end.

feat(): Added --exit-sucess option
This option ensures that the program exists successful, i.e., exit(0) is
called, even in the event of finding unused variabels. The option is
meant to be used with the 'sonar' --formatter type, if a sass linter is
being run after fusv, so that reports of fusv and the sass linter can
report to sonarqube together.

feat(parse-variable.js|index.js): calc ending line/column of vars
refactor(cli.js): created own function for printing results
feat(cli.js): Added option --output-file for creating a report file

refactor(cli.js): renamed unusedList to unusedVarsReport

moved printing to console functionality to report.js

createReportString() uses ansi codes for console and skippes them for
writing to a report file.